requirements.txt
python-dotenv==1.0.1
homeassistant==2024.3.3
requests==2.31.0